% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{od_data}
\alias{od_data}
\title{origin-destination matrix}
\format{
A datatable with 16384 rows and 2 variables
\describe{
\item{Name}{trip id; structure: 'origin_id - destination_id'}
\item{travel_time}{travel time by car in minutes between origin and destination}
}
}
\source{
\url{https://marcinstepniak.eu/project/giservice/}
}
\usage{
data(od_data)
}
\description{
Origin-destination matrix of travel times by car between all census tracks of Ursynów district in Warsaw.
}
\keyword{datasets}
