% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rej_ursynow}
\alias{rej_ursynow}
\title{Polygons of census tracks of Ursynów district}
\format{
A spatial dataset with 128 rows and 3 variables
\describe{
\item{id}{official id of a census track}
\item{POP}{total population of a census track}
\item{geometry}{geometry of a census track}
}
}
\source{
\url{http://www.codgik.gov.pl/index.php/darmowe-dane/prg.html}
}
\usage{
data(rej_ursynow)
}
\description{
Spatial dataset of 128 census tracks of Ursynów district of Warsaw
}
\keyword{datasets}
